requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
python-dotenv==1.0.1
langchain==0.2.12
langchain-openai==0.1.22
langchain-community==0.2.11
faiss-cpu==1.8.0.post1
rapidfuzz==3.9.6
tiktoken==0.7.0
streamlit==1.37.1
redis==5.0.7
langwatch==0.2.9

